requirements.txt
streamlit
pypdf
python-docx
sentence-transformers
scikit-learn
langchain_community
langchain_core
langchain_groq
faiss-cpu
